move LudoboxError to its own file
