+++ tests/test_content.py
@@ -5,13 +5,19 @@ import os
 import unittest
 
 from ludobox.config import read_config
-from ludobox.data.content import build_index
+from ludobox.content import build_index, read_game_info
 
 class TestLudoboxContent(unittest.TestCase):
     """Functions to index, sort and search content"""
 
     def setUp(self):
         self.config = read_config()
+
+    def test_read_game_info(arg):
+        """Make sure an info file is read properly"""
+        path = os.path.join(os.getcwd(), 'tests/test-data/test-game/info.json')
+        read_game_info(path)
+        self.assertTrue(False)
 
     def test_build_index(self):
         """Should build an index containing all data in this field"""
